Remove Test::Synopsis plugin
There are a _lot_ of issues with the synopses in this distro. It'd be nice to
fix this later, but it also might require being able to selectively skip some
modules. For example the Test2::Require::EnvVar synopsis causes the test suite
to bail because it executes at compile time.